PoneglyphMiddleware/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
pika==1.3.2
redis==5.0.1
boto3==1.34.0
python-dotenv==1.0.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
prometheus-client==0.19.0
structlog==23.2.0
asyncio-mqtt==0.16.1
